requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dotenv==0.19.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
easyocr==1.7.1
PyMySQL==1.1.0
cryptography==44.0.1
fastapi==0.110.0
uvicorn==0.27.1
mysqlclient==2.2.0
pillow==10.3.0
python-multipart==0.0.19